sets dashboard to loading state while waiting for user
