+++ src/components/Dashboard.js
@@ -1,9 +1,10 @@
-import React from "react"
-import { connect } from "react-redux"
+import React, {useEffect} from "react"
+import { connect , useDispatch} from "react-redux"
 import { Route, Switch } from "react-router-dom"
 
 // Functions
-import { toggleStep, fetchRequirements } from "../actions"
+import { toggleStep, fetchRequirements, fetchUser } from "../actions"
+import { loadAuthDataFromLocalStorage } from "../store"
 
 // Components
 import UserInfo from "./UserInfo"
@@ -11,12 +12,22 @@ import RequirementCardContainer from "./requirement/RequirementCardContainer"
 import Navigation from "./Navigation"
 import RequirementDetails from "./RequirementDetails"
 import NotFound from "./NotFound"
-
+import FullPageLoader from "./lib/Loaders/fullPageLoader";
 const Dashboard = props => {
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    const { userId, token } = loadAuthDataFromLocalStorage()
+    dispatch(fetchUser(token, userId))
+  }, [dispatch])
+
+  if (!Object.keys(props.user).length ) {
+    return <FullPageLoader />
+  }
   return (
     <div className="dash-container">
-      <Navigation />
-      <UserInfo />
+      <Navigation user={props.user}/>
+      <UserInfo user={props.user}/>
       <Switch>
         <Route
           exact
@@ -32,12 +43,16 @@ const Dashboard = props => {
 
 const mapStateToProps = state => {
   return {
-    inProgress: state.stepReducer.inProgress,
-    stepsByTask: state.stepReducer.stepsByTask
+    user: state.userReducer.user,
+    requirements: state.requirementReducer.requirements,
+    stepsInProgress: state.stepReducer.inProgress,
+    stepsByTask: state.stepReducer.stepsByTask,
+    requirementsInProgress: state.requirementReducer.inProgress,
+    userInProgress: state.userReducer.inProgress 
   }
 }
 
 export default connect(
   mapStateToProps,
-  { toggleStep, fetchRequirements }
+  { toggleStep, fetchRequirements, fetchUser }
 )(Dashboard)
