+++ src/components/Navigation/index.js
@@ -11,13 +11,10 @@ import Dropdown from "../lib/Dropdown"
 import Flyout from "../lib/Dropdown/Flyout"
 import {history} from "../../store"
 
-const Navigation = props => {
-  const [user, setUser] = useState({ first_name: "Loading..", last_name: "" })
+const Navigation = ({user}) => {
   const [width, setWidth] = useState(window.innerWidth)
   const [isOpen, setIsOpen] = useState(false)
   useEffect(() => {
-    const { token, userId } = loadAuthDataFromLocalStorage()
-    getUserById(token, userId).then(user => setUser(user))
     const handleResize = () => {
       return setWidth(window.innerWidth)
     }
